- Add ability to set programs to allow keyboard controls while running. (meant only for inference-only programs that don't press buttons) - Add in-progress work on new shiny detection.

+++ SerialPrograms/Source/CommonFramework/ImageTools/DistanceToLine.h
@@ -18,25 +18,25 @@ namespace PokemonAutomation{
 class DistanceToLine{
 public:
     DistanceToLine(
-        double vertex0_x, double vertex0_y,
-        double vertex1_x, double vertex1_y
+        ptrdiff_t vertex0_x, ptrdiff_t vertex0_y,
+        ptrdiff_t vertex1_x, ptrdiff_t vertex1_y
     )
         : m_vertical(vertex0_x == vertex1_x)
         , m_vertex0_x(vertex0_x)
         , m_vertex0_y(vertex0_y)
 //        , m_vertex1_x(vertex1_x)
 //        , m_vertex1_y(vertex1_y)
-        , m_slope((vertex1_y - vertex0_y) / (vertex1_x - vertex0_x))
+        , m_slope((double)(vertex1_y - vertex0_y) / (vertex1_x - vertex0_x))
         , m_inverse(1 / (m_slope*m_slope + 1))
     {}
 
-    double distance_squared(double x, double y) const{
+    double distance_squared(ptrdiff_t x, ptrdiff_t y) const{
         if (m_vertical){
-            double distance = x - m_vertex0_x;
-            return distance * distance;
+            ptrdiff_t distance = x - m_vertex0_x;
+            return (double)(distance * distance);
         }
-        double dx = x - m_vertex0_x;
-        double dy = y - m_vertex0_y;
+        ptrdiff_t dx = x - m_vertex0_x;
+        ptrdiff_t dy = y - m_vertex0_y;
         double top = m_slope * dx - dy;
         double distance_sqr = top * top * m_inverse;
         return distance_sqr;
@@ -44,10 +44,10 @@ public:
 
 private:
     bool m_vertical;
-    double m_vertex0_x;
-    double m_vertex0_y;
-//    double m_vertex1_x;
-//    double m_vertex1_y;
+    ptrdiff_t m_vertex0_x;
+    ptrdiff_t m_vertex0_y;
+//    ptrdiff_t m_vertex1_x;
+//    ptrdiff_t m_vertex1_y;
     double m_slope;
     double m_inverse;
 };
